Move to addonStoreGui.controls.messageDialogs
